Move locally built plugins handling up to AbstractKotlinIntegrationTest
